+++ CHANGES.rst
@@ -69,6 +69,7 @@ Improvements
   :user:`Michi03`)
 - Include abstract details in comment notification email subject (:issue:`6449`, :pr:`6782`,
   thanks :user:`amCap1712`)
+- Use markdown editor field in survey questionnaire setup (:pr:`6783`, thanks :user:`amCap1712`)
 
 Bugfixes
 ^^^^^^^^